Testing out a guessing game
enter the parameters of what number you want to get a high and low number and you get 7 guesses to find out if you are correct. After the 7 attempts the result if wrong prints the correct answer. The guesses do include the two numbers that were input for the starting parameters.